Move AddSystem to inputs
